Problem: (Fix #967) MultiSig logic is coupled with MultiSig service
Solution: Decouple MultiSig into separate module for re-use
