+++ client-core/src/multi_sig.rs
@@ -1,0 +1,8 @@
+//! MultiSig operations support
+mod builder;
+mod session;
+mod signer;
+
+pub use builder::MultiSigBuilder;
+pub use session::MultiSigSession;
+use signer::Signer;
